♻️ refactor: extract agent execution types into dedicated agentExecution module
